Use custom paste action instead of overriding the native one. (#1116)
