+++ package.json
@@ -394,12 +394,6 @@
           "default": true,
           "description": "Enable/disable Smart Selection support for Java. Disabling this option will not affect the VS Code built-in word-based and bracket-based smart selection.",
           "scope": "window"
-        },
-        "java.actionsOnPaste.organizeImports": {
-          "type": "boolean",
-          "description": "[Experimental] Triggers 'Organize imports' when code is pasted into a Java file. This overrides VS Code's default paste action, thus might cause some issues. Disabled by default.",
-          "default": false,
-          "scope": "window"
         }
       }
     },
@@ -469,6 +463,12 @@
       {
         "command": "java.workspace.compile",
         "key": "shift+alt+b"
+      },
+      {
+        "command": "java.action.clipboardPasteAction",
+        "key": "ctrl+shift+v",
+        "mac": "cmd+shift+v",
+        "when": "javaLSReady && editorLangId == java"
       }
     ],
     "menus": {
